docs: remove navbar examples and general example
