--- example/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "example",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "scripts": {
-    "start": "parcel index.html",
-    "build": "parcel build index.html"
-  },
-  "dependencies": {
-    "bootstrap-italia": "^2.7.10",
-    "react-app-polyfill": "^1.0.0",
-    "typeface-lora": "^1.1.13",
-    "typeface-roboto-mono": "^1.1.13",
-    "typeface-titillium-web": "^1.1.13"
-  },
-  "alias": {
-    "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom/profiling",
-    "scheduler/tracing": "../node_modules/scheduler/tracing-profiling"
-  },
-  "devDependencies": {
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.8.4",
-    "parcel": "2.10.3",
-    "typescript": "^4.2.4"
-  }
-}
