+++ src/index.tsx
@@ -106,7 +106,6 @@ export { TextArea } from './Input/TextArea';
 export { GoBack } from './GoBack/GoBack';
 export { LinkList } from './LinkList/LinkList';
 export { LinkListItem } from './LinkList/LinkListItem';
-export { Megamenu } from './Megamenu/Megamenu';
 export { MegamenuItem } from './Megamenu/MegamenuItem';
 export { MegamenuFooter } from './Megamenu/MegamenuFooter';
 export { MegamenuHighlightColumn } from './Megamenu/MegamenuHighlightColumn';
@@ -207,7 +206,6 @@ export type { InputContainerProps } from './Input/InputContainer';
 export type { TextAreaProps } from './Input/TextArea';
 export type { LinkListProps } from './LinkList/LinkList';
 export type { LinkListItemProps } from './LinkList/LinkListItem';
-export type { MegamenuProps } from './Megamenu/Megamenu';
 export type { MegamenuItemProps } from './Megamenu/MegamenuItem';
 export type { MegamenuFooterProps } from './Megamenu/MegamenuFooter';
 export type { MegamenuHCProps } from './Megamenu/MegamenuHighlightColumn';
