+++ stories/Megamenu/Megamenu.stories.mdx
@@ -1,6 +1,6 @@
 import { ArgsTable, Meta, Canvas, Story } from '@storybook/addon-docs';
 
-import { Megamenu, MegamenuItem, MegamenuFooter, MegamenuHighlightColumn } from '../../src';
+import { MegamenuItem, MegamenuFooter, MegamenuHighlightColumn } from '../../src';
 
 <Meta title='Documentazione/Menu di navigazione/Megamenu' of={MegamenuItem} />
 
@@ -66,12 +66,6 @@ Per aggiungere un elenco di _call to action_ posizionato a destra, inseriremo ta
 </Canvas>
 
 ## Argomenti componente
-
-### Megamenu
-
-È un wrapper che utilizza come elemento principale `Navbar`. Ne accetta quindi tutti gli attributi.
-
-<ArgsTable of={Megamenu} />
 
 ### MegamenuItem
